Unbind Mac option key
The Mac option key is critical for inserting symbols like curly braces
(on a German keyboard layout) or German umlauts (on a US keyboard
layout). So the standard Emacs way of binding it to Meta makes these
symbols unaccessible.

Instead, I am now using the left Mac command key as Meta and the right
mac command key as Super.